fix(triton): widen blocked RMSNorm row index (#1482)
## Summary

Widen the blocked RMSNorm forward program ID before row-offset address
arithmetic. The default blocked path (`n_cols <= 256`, `n_rows >=
32768`) otherwise overflows a 32-bit row address at 2^31 elements.

This completes the same safeguard introduced for the non-blocked RMSNorm
kernels in #804.

Fixes #1483.

## Testing Done

- Hardware Type: NVIDIA RTX 2000 Ada Generation Laptop GPU (8 GB)
- [x] `python -m pytest -q test/transformers/test_rms_norm.py -k
block_rms_norm -x` (1 passed; large CUDA regression skipped because it
requires 8.6 GB free VRAM)
- [x] `ruff check src/liger_kernel/ops/rms_norm.py
test/transformers/test_rms_norm.py`
- [x] `ruff format --check src/liger_kernel/ops/rms_norm.py
test/transformers/test_rms_norm.py`
- [ ] `make test` (make is unavailable on this host)
- [ ] `make checkstyle` (make is unavailable on this host; its ruff
commands above passed)
- [ ] `make test-convergence` (not run)

+++ src/liger_kernel/ops/rms_norm.py
@@ -272,7 +272,7 @@ def _block_rms_norm_forward_kernel(
     3. https://arxiv.org/pdf/1910.07467
     """
 
-    row_idx = tl.program_id(0) * BLOCK_ROW + tl.arange(0, BLOCK_ROW)
+    row_idx = tl.program_id(0).to(tl.int64) * BLOCK_ROW + tl.arange(0, BLOCK_ROW)
     col_offsets = tl.arange(0, BLOCK_SIZE)
     row_mask = row_idx < n_rows
     col_mask = col_offsets < n_cols
